convert lots of doubles to floats

+++ inc/FMTranscoder.hpp
@@ -12,7 +12,7 @@ private:
     uint32_t basebandFrequency;
     bool predeEmphasis;
     bool limitDeltaPhase;
-    double preEmphasisFilter[2];
+    float preEmphasisFilter[2];
 
 public:
     /**
@@ -37,10 +37,10 @@ public:
 
         // Initialize pre-emphasis filter taps
         preEmphasisFilter[0] = 1.0;
-        const double timeConstCutoff = 50e-6; // τ = 50µs in Europe
+        const float timeConstCutoff = 50e-6; // τ = 50µs in Europe
         preEmphasisFilter[1] = -(1.0 - 1.0 / (timeConstCutoff * samplingFrequency));
         centralFrequency = (2.0 * maxModulatingFrequency) / samplingFrequency;
     };
 
-    void modulate(const double *signal, uint16_t signal_length, double *in_phase_signal, double *quadrature_signal);
+    void modulate(const float *signal, uint16_t signal_length, float *in_phase_signal, float *quadrature_signal);
 };
