+++ inc/filter.hpp
@@ -10,13 +10,13 @@ constexpr uint8_t quadratureSync = 0b01;
 // sure to make use of the existing HAL library functions (etc
 // `LL_FMAC_FUNC_CONVO_FIR`)
 
-void filterFIR(const double *filterTaps, uint16_t numberOfTaps,
-               const double *inputSignal, uint16_t size, double *outputSignal);
+void filterFIR(const float *filterTaps, uint16_t numberOfTaps,
+               const float *inputSignal, uint16_t size, float *outputSignal);
 
-void integrate(const double *inputSignal, uint16_t size, uint16_t numberOfTaps,
-               double *outputSignal);
+void integrate(const float *inputSignal, uint16_t size, uint16_t numberOfTaps,
+               float *outputSignal);
 
-void multiplyVector(double *signal, uint16_t length, double coef);
+void multiplyVector(float *signal, uint16_t length, float coef);
 
 void packetizeIQ(double *inPhaseSignal, double *quadratureSignal,
                  uint16_t length, uint32_t *outputSignal);
